Add XUP Menu Traversal utility
Utilize the "List Options by Parents and Use" option to create two option
dictionaries that can be used to simplify the menu traversal
process. One dictionary is of the option linked to a parent option,
the other is the option's internal name linked to its Menu Text,e.g.
EVE -> Systems Manager Menu

Add a function to the OSEHRAHelper which uses these new dictionaries to go
to an XUP menu from either the prompt or an XUP Menu.  When that function is
called, go through the dictionary to find a path from that option to one
that has no "parent".  This becomes the menu path to the option.

This new function takes 2 required arguments with 1 optional:

goToOption(<internal menu name>,<binary dir path>,[<access code>])

As the menu is accessed, update a parameter on the connection that keeps
track of the current menu location.  This array is how common menu paths
are calculated for moving from one menu to another.  Another parameter
added is the lastconnection, which captures the text that is searched by
the various "wait" functions.

Add a test file that compares the menu path for one menu to a
predetermined baseline and visits three different menus and add a "Smoke
Test" to attempt to visit all menus in the instance from the top down.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-52
Change-Id: I2a1942f012184e3cabc1d1fc95da767d75d867b6