Problem List: Fix ZZRGUTRB unit test
HISTORY^ZZRGUTRB fails on certain conditions. This is a timing issue
related to the way Problem List audit log events are indexed. If many
events happen in a short interval (less than a second) the reverse order
index is not deterministic.
So now HISTORY checks for the correct audit events without assuming a
specific order.

Change-Id: Ibba9839e2f4358662cb3093c2639ca5c6d40d7bf