Bug: Fix incorrect exception path being checked
The Package name is being passed to the ReportXINDEXResult functions have
underscores instead of spaces. If the underscores are not replaced, the
exceptions for the file will not be found. As is, the exception file for
the first error encountered will not be found due to the incorrect package
name.

Change the location of the replace command
to not just be after an exception was not found.  Instead place it just
before the foreach that searches the output.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-73
Change-Id: Idb22f4a507edee3ebc84bd4aea0e483cbb83ad7c