Update rem_all in PLActions.py to eliminate paging
Change the self.VistA.write('??') command at the end of the rem_all
function in PLActions.py to send '?' instead.  The '??' will display
options for the user's secondary menu, which could lead to it paging
depending on the amount of menus that the user possesses.

The single '?' will only show the menu's options and shouldn't require any
other interaction

OSEHRA-Id: http://issues.osehra.org/browse/OAT-79
Change-Id: Ic5c5fe2a9f70c6fffc68a0657da5a4737a6a2870