PatchSequenceApply: Warn and ignore missing info-only dependencies
A missing dependency named only in the patch info file should not
prevent a KIDS build from being installed if that extra dependency is
not found in INSTALL file of a VistA instance.  Issue a warning instead.

This is necessary to install patch "GMPL*2.0*43" whose info file adds an
extra dependency on "GMPL*2.0*5".  The dependency is too old and no
information related to this patch can be found in the FOIA VistA INSTALL
file.

Change-Id: I81ad75a625dbf936c341ac6bbde5030b53fee064