PatchOrderGenerator: Include isolated patches for topological sort
Also include patches that do not have KIDS build for topological sort,
but remove them from the final patch order generated.

Add logic to remove patches that are not supposed to be installed
according to csv file but somehow we have the KIDS build in repository.

Add check to report out of sequence order error for multibuilds.

TestPatchOrderGenerator: Update testing case to include an
isolated node for toplogic sort.

OSEHRA-Id: http://issues.osehra.org/browse/VAF-140
Change-Id: I0fca022390fde03150ed730f803a85e45a70f9cb