PatchIncInstallEx: Fix a bug with backup in continuous mode.
If backup option is set and runs in continuous mode, for each backup
it will have to shutdown cache instance and that make test client previous
created invalid. The fix is to create a new test client for each patch run.

Also set the VistA repository dir to be script directory in
generateCommitMsgFileByPatchInfo api call so that the script can be invoked
from anywhere outside VistA repository.

Change-Id: Icbc79502b5babfeea374d315908a21cb7fcfade1