Add option for Custom Refresh command
The current importRG framework assumes a very basic configuration for
the VistA instances.  Add support for the user to add a custom command to
be used in place of the current database delete/rebuild steps.

The option "TEST_VISTA_FRESH_CUSTOM_REFRESH" is marked as advanced as is
the corresponding string variable "TEST_VISTA_FRESH_CUSTOM_COMMAND".

The COMMAND variable will need to contain both the executable in quotation
marks and the path to the file to execute:
"sh" ~/refresh.sh   or "/bin/python" ~/refresh.py

OSEHRA-Id: http://issues.osehra.org/browse/OAT-81
Change-Id: I757f5ef2ee10a5353be5ebf65eb7350b74f9871b