Update Sikuli script handling of RAS information
The Sikuli script expect the "VISTA HEALTH CARE" string to be visible in
the list of clinics when creating a visit during the test.  The new
clinics from the Scheduling RAS tests push it off the visible list.

Update the commands to no longer expect to click on it, but type the name
of the clinic into the "Visit Location" box above the list.  Remove the
old screenshot.

The problem categories also changes the behavior of CPRS.  Update the
script to click on the "Other Problem" to pop up the problem search box
that we need to interact with.

Change-Id: I29786c56d9f5345ca70d998613066a81466d31dc
OSEHRA-Id: http://issues.osehra.org/browse/OAT-84