EHN: Fix XINDEX Testing error on GT. M
In patch XT*7.3*133, a new tag OBJTST in XINDX2.m is added with
code specific for Caché, and this change causes XINDEX testing to stop
working for some of packages on GT.M, as it throws an error for any
routines that contains Caché specific code that starts with '##'.

Modify OBJTST tag in XINDX2.m to return 1 if not on Caché, also make
change to import the modified version of XINDX2 during the setup process.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-95
Change-Id: I879f1fe6aa2b82223863c921067512077e8e4664