Fix environment variables, Add p and s directories
The gtmroutines variable was coded in two places for x64 and x32 -
factor out the common parts to another variable that is included in the
x64 and x32 if statement

The gtm_zinterrupt variable contained characters that were interpreted
by the bash script and not included as intended in the env script

Make certain variables available to other scripts via $basedir/etc/env
so they can be chained together easier.

The p and s directories are for primary and secondary development
respectively. These directories are useful for installing nodem and
other hot fixes. These must be added after VistA is installed so
ZTMGRSET doesn't get confused.

OSEHRA-Id: http://issues.osehra.org/browse/VAF-141
OSEHRA-Id: http://issues.osehra.org/browse/VAF-142
OSEHRA-Id: http://issues.osehra.org/browse/VAF-144
Change-Id: Ic31f0e0b6eeb160a784efc2408bce1ac18a19e0f