Update rST Documentation images
Match the images in SetupTestingEnvironment.rst to the current state of
the repository.  The text has been updated, but not the images.

Shrink the amount of extra content in the images to facilitate the
replacement in the future by taking the screenshot of only the desired
content.

Various grammatical updates and limit paragraph line length to
80 characters.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-97
Change-Id: Ia01b6086b867d934a514f124949b63de82658a38