Add VA VistA patches from Jul, Aug, Sep 2013
Also add CPRS 29 related patches from June 2013 and patch XU*8*614.

Follow the instructions in 'Scripts/HowtoDownloadPatches.rst' to
download patches from the VA FOIA site.  Use the following shell code to
produce the URL list and pipe it to wget:
{
for f in \
 "2013/July 2013" \
 "2013/August 2013" \
 "2013/September 2013" \
 "2013/June 2013/CPRS VERSION 29 - ALSO IN INDIVIDUAL APPS" \
; do
  echo "https://downloads.va.gov/files/FOIA/Software/Patches by Year and Month Released/$f/"
done
}

{
for f in \
 "XU-KERNEL/XU-8_SEQ-497_PAT-614.TXT" \
 "XU-KERNEL/XU-8_SEQ-497_PAT-614.KID" \
; do
 echo "https://downloads.va.gov/files/FOIA/Software/Patches_By_Application/$f"
done
}

Retrieve .TXT/.KID files posted as of 2013-11-19.

OSEHRA-Id: http://issues.osehra.org/browse/VAF-154
Change-Id: I78e2770d0121a737cb687f0b3d6372a1d5e9f0c7