Prepare Tests for FileMan 22.2 submission
The testing of Fileman 22.2 highlighted some eventual problems with the
RAS tests, fix the commands to account for the changes that will come.

Update the wait_re function to compile the supplied string argument into a
case insensitive expression to be used in the match.

Prepare for:

The changing of "Press RETURN" to "Press Enter" in prompts by removing the
references to those keys and waiting for "to continue".
Changing of phrase used to exit by using a multiwait to capture both
possible phrases
Changing of case of upper to mixed in prompts by using the new
case-insensitive wait_re functions for certain commands.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-108
Change-Id: Ief745019a4e9a95979fb2beab359813221806191