Add Exceptions for VA Fileman 22.2 routines
Add XINDEX exceptions for the new routines that are added in the
FileMan22.2 submission.

DINVGTM/DINVGUX: New routines that contain logic specific to GT.M on VMS
and Unix.  Routines similar to this for other platforms were exempted and
the error found in both routines has been exempted in Kernel.

DMUD*: MUnit Test routines, should not be run in normal operation.  If
using the OSEHRA framework, the MUnit routines will be installed prior to
running these tests.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-108
Change-Id: I925525a5dd560c47f85612eb0a3af7f1eb6ba6cd