Fix MUnit error parsing
Fix the CommonFunctions.cmake code that parses the MUnit output for error
reporting.  It assumed that the all MUnit routines would start with ZZUT which
is not the case.

Remove the "ZZUT" from the regex matches while changing the failure
message line to check for the captured routine_name.

Add comments to describe each if statment.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-112
Change-Id: I1f71123b83825e6e47bc7659be42936409671393