Fix line endings part 2 & Fix xinetd.d scripts
If a repository is already cloned and a user does a 'git pull' to update
the repository on a windows machine the line endings for the shell
scripts will be incorrect. This change will force line endings to be
line feeds and git shouldn't complain about the change due to the
.gitattributes in a previous commit.

In a previous commit which allowed changeable instance names a couple of
files were missed. This implements a quick fix to search and replace the
hard coded default of foia to $instance.

Change-Id: If69d77db3f4c642ceec44a7cd2fa1d4119e0c006
OSEHRA-Id: http://issues.osehra.org/browse/VAF-162
OSEHRA-Id: http://issues.osehra.org/browse/VAF-163