Change MUnit command generation code
Change the CMake code that generates the command list for the MUnit tests.
The current setup assumes that the entire routine will be devoted to
testing, while there is the potential of an separate entry point to run
the testing.

Remove the "D ^" from the command generation and rely on the line from
TestList to have one correctly. Read a command from the TestList file instead
of just the name of the routine to test.

IE, in Registration's TestList: Turn  "ZZDGPTCO1"  into  "D ^ZZDGPTCO1"

Update the REGEX in the CommonFunctions.cmake that finds the routine name
from the output to account for a potential entry tag.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-117
Change-Id: I9996d08e6aa7778fcff9789f1999e5381144feea