Modularize and comment PostImportSetupScript.py.in
Simplify the PostImportSetupScript by creating functions in a new class
called OSEHRASetup to perform the same steps as the original file.

Comment each function to explain the process and what is being created.

Change the references to the Manager,System in the RAS testing to use
correct form of the Verify Code after changing.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-119
Change-Id: I8a8cb260307a319e4468cc4bcef4ca64d906f79b