Add UserTest directory for integration
Add a UserTest directory that contains only a CMakeLists.txt file to help
add tests to the environment without understanding the complexities of the
repository.

Add a blank CMakeLists.txt to the UserTest directory with a message
explaining the purpose of the file with an example command.

Add an option "TEST_VISTA_USERTEST" to control whether to run the
add_subdirectory command.

Add an add_subdirectory command for the UserTest directory the command
should only be run if the TEST_VISTA_USERTEST option has been selected.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-122
Change-Id: Iadcc11467818f5ec6a52b788ce3f328ee7fb6725