Add "Verify Fields" test for Fileman Data entries
Add a test that uses the "Verify Fields" utility of Fileman to check the
files associated with each package.

It should get the list of files to check for each package from the
"File Numbers" field in the Packages.csv and enter each file by selecting
the file by IEN:  `####.

Add a VerifyFields python file that will be configured with connection
information and will access the FileMan "Verify Fields" utility for each
File found in the Packages.csv.  It should stop the display of errors
after the first error message is reported.

Due to timeouts occuring with exceptionally large files, add steps
in the script to skip checking files that have >300,000 entries.

Add an option in Testing/CMakeLists.txt to control whether to run the
VerifyFields tests, and mark the option as advanced.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-126
Change-Id: Idbf5860a63c101e616fdd57c8476931fd9b2d874