Fix bug in connection to password protected instances
A set of commands to the Connect*Cache functions is incompatible with
instances that require a username/password to access the programmer
prompt.  It was added to attempt to  ZN into a namespace before running
the test script.  It is not necessary and causes an issue.

Remove the commands and add them back to the test
functions that are missing them.

Change-Id: I7a4183a95c931128ee63c9e8d9b6d81615ec0c4f
OSEHRA-Id: http://issues.osehra.org/browse/OAT-123