RPCBrokerCheck arguments instead of configuration.
Replacing the process of configuring the RPCBrokerCheck.py.in file, with a
mechanism for passing arguments in the command line from the CMakeLists.txt
file, to the TestSSEP.py file, and ultimately to the actual function in
RPCBrokerCheck.

This avoids the generation of a RPCBrokerCheck.py file in the source tree.

Change-Id: Ic39e467def18de3ad0e9fd0e7969f9dca66789e7