Change Sikuli test to prepare for CPRS V29
Change the problem creation code to prepare for the different process that
is required to create a problem with the 799.9 ICD code.  CPRS v29 has
more steps to enter that data.

Change the step after entering the signature code for the strep test and
the first signature to hit "Enter", rather than clicking on the "OK"
button, as CPRSv29 changes the button to read "Sign"

Change the images used to select the problems.  CPRSv29 doesn't display
the ICD code with the name in the search parameter, it resides in a
separate column.  Take a new screenshot with only the Problem's string.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-131
Change-Id: I87f122a127cff84d9b2c95b889821ae0142749ee