Fix username/password bug for RAS tests
The code used in the OSEHRA RAS functional tests are unable to
run on an instance that uses a username/password prompt before displaying
the MUMPS prompt.

Change the RASTest.cmake.in to export the VISTA_CACHE_USERNAME and
VISTA_CACHE_PASSWORD value to environment variables.

Change the TestHelper code that connects to the instance to attempt to grab
the VISTA_CACHE_USERNAME value.  After connection, the
test_suite_details.username will be checked for a value.  If it exists,
the code should grab the VISTA_CACHE_PASSWORD value and use them to
answer the sign in prompts.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-132
Change-Id: I2e927f6b9447b1f744d866470491859dcaf7a51f