Fix if statements in Registration RAS device handling
The Registration test fails in Linux/Cache environments due to code that
has the script expecting to see the "Right Margin" prompt after selecting
the device to write to.  This prompt has only been seen on Windows
evironments.

Change the 'if' statments that control this code to use

" sys.platform == 'win32' "

instead of

" self.VistA.type == 'cache' "

Import the sys library where necessary.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-133
Change-Id: I534e471214c2e6d027c0beaffd21b52ace7803b4