Add Lab Service Execeptions
Add a set of exceptions in the Lab Services package where calls are being
made into the Imaging package.  The errors expressed
here are found on lines where the routine existence is being checked.  The
errors were found in the March 2014 FOIA release.

Checking the return of the $TEXT function acts as a code guard to ensure
that the calls to the Imaging package will not be executed if the routine
isn't present.

OSEHRA-Id: http://issues.osehra.org/browse/VAF-169
Change-Id: I44530fc1405378148d55bcaff4ae7fbe8e1279df