Fix RAS tests to adjust to Mar 2014 FOIA
With the VA removing the programmer access for the DUZ=1 user, we call the
fileman entry point (D Q^DI) after setting the DUZ to ensure that the
DUZ(0) is set to '@'.

The name of the DUZ=1 user has also changed, update the Scheduling test to
not check for "TESTMASTER" but for "USER"

Update the Problem List RAS tests to continue to put the ICD9 code to
select a problem, when it doesn work, write in a SNOMED code.

Adjust for the absence of the "Veteran ID Card Menu" from the Menu
system.  The menu was removed in patch DG*5.3*874.  Adjust the script to
only attempt to access the menu if it is found in the text of the
avialable options of the ADT Managers Menu.

OSEHRA-Id: http://issues.osehra.org/browse/VAF-171
Change-Id: Ib51cc2ccca5c427edaa5d9786836d7fd59ada3e7