Fix Problem List MUnit test for Mar 2014 FOIA
With the SNOMED code additions, the stored names of the Problems has been
changed to have the ICD9 code in the string.  Adjust the problem list to
check the result for everything but that additional ICD9 string.

Also, change references to the TESTMASTER,USER or other users names to be
more generic, due to the renaming of certain users in a recent FOIA
release.

OSEHRA-Id: http://issues.osehra.org/browse/VAF-171
Change-Id: Idbffc951d4d5ba0d3f2b05c26916057850635b6e