Add test to pull from instance for Coverage Calc
Add a test and python file to export/unpack all routines from the VistA
instance during testing.  Change the path in the *mcov files to use this
directory as the source for the coverage calculations.

This ensures that the routines being run during the test are used as the
source instead of the VistA-M repository.

Move the coverage generation code below the MUnit test generation.  This
will allow a list of MUnit tests to be generated and passed as a
dependency to the new routine export test.  This ensures that all testing
routines will be counted in the calcuations.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-139
Change-Id: I73378e50cf65f4b1366baf27eb0e8359435765ea