Fix Testing/CMakeLists.txt ordering
Changing the order of the Testing/CMakeLists.txt code caused some problems
with the testing harness.

Putting the MUnit code before the Coverage Calculation to get the
dependency list caused the MUnit tests to not be configured with the
correct routine subset.

The move also allowed the MUnit routines to be checked during the XINDEX
tests which caused some extra errors.

Reorder the code to have the order be : coverage, XINDEX, MUnit.  Remove the
the dependency on the MUnit tests.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-144
Change-Id: I656932e299d8849dd5c991ce65f780bc9d4000b7