Add support for nodem 0.3.3
nodem version 0.3.3 changed the name of the file for GT.M call-in
support (calltab.ci => nodem.ci) and the name of the supporting M
routine (node.m => v4wNode.m).

Modify ewdjs.sh to dynamically determine the name and path of the .ci
file. The script already knows how to deal with the .m file.

Change-Id: I06d2327fdfe199a3d7428ac2e740640bdbb3de34
OSEHRA-Id: http://issues.osehra.org/browse/VAF-185