Fix Registration RAS test prompt
Fix the Registration's RAS test script.  A change is necessary when the
adt_smoke_test call of the "Summary of Dispositions" option asks for a
start date. if the start date falls on the first of the month, there will
be an extra prompt to ask if it wants statistics for the whole month that
the script doesn't account for.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-145
Change-Id: I4edbbd35f179c9ee6c69fed3d8d430ebd0b97f6b