Update Testing to account for updates
Update the testing code to account for functional changes in the VistA
codebase in the November 2014 FOIA release.

Update prompts in the Registration and Scheduling Roll-and-Scroll tests

Update the information given to the Problem List MUnit test and adjust
what is checked for on depending on the contents of the output.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-160
Change-Id: Ib3ed1db6a5f8e5f03f89719bc62998bd061cc7f7