Add prettify files for future Dox applications
Add the prettify files needed for syntax highlighting in the source code
pages of the Dox tool.  Since the files are licensed with the Apache 2.0
license, we can simply place the files we need into the proper directory

https://github.com/google/code-prettify

Add instructions on where to place the folder after generating the web
pages to ensure that the files are found.

OSEHRA-Id: http://issues.osehra.org/browse/VIVIAN-112
Change-Id: Id16ba2ab6b24f7e8bb9484e5d839443622fbb717