Fix UNITTEST for the Problem List Package
Add the ICD10 code used for "Diabetes Insipidus" of "R69." to allow the
problem check to be returned successfully.
Change from the TRM being 1 of unknown to being the SNOMED code for
Essential Hypertension.

Change the ZZRGUTEX routine to use the ICD10 Code for hypopituitarism and
change the coding system to correspond to the new code

The ZZRGUTRB routine will continue to fail due to an issue which can be
read about here:  http://issues.osehra.org/browse/CONV-111

OSEHRA-Id: http://issues.osehra.org/browse/OAT-178
Change-Id: I30b9fba4ef46e19052296174bd767a58d9ddead3