Clarify Dox setup documentation
Fix the documentation about the generation of the Dox pages.  Add more
information about the M Routine Analyzer, where to run commands and more
information about the necessary environment variables.

Add a note to the -o of the WebPageGenerator that the directory needs to
exist before the running of the script.

change from `hightlighting` to ``highlighting``.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-180
Change-Id: I958374ad866979346aa8d4faf98d21bb0bbb869c