Add custom installer for M-Unit.
To answer the routine copy questions which appear on the GT.M system during
the run of the post install routine, a custom installer is needed for the
automatic install.

Add the necessary script.  The testing ran into a timeout when rebuilding the
menus after install, so the timeout for the exitKIDSMenu was increased.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-183
Change-Id: I3a87361d46c4d6516552ddc33bfad8b0b5e637ba