Change the OSEHRA Patient setup
Add a new .ro file to be installed during the setup which eliminates the
calls to the entry points where the Cache Classes are called. These
changes were made in the DPTLK7 routine.  Add the code to install this .ro
file during the run of the ImportRG script.

Change the existing .ro files to be .gtm.ro to continue to be installed
only on the GT.M systems.

Add the code to account for the new Enterprise Search as part of the
patient registration code in OSEHRASetup. Also, combine the two current methods
of adding patients to the VistA instance by reading the patdata0.csv file
as part of the PostImportSetupScript.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-189
Change-Id: I5d3dd05dc61405ac7269ba8d9c723dbb5265a7d9