Add file to update the MFS globals
Add a script which downloads the globals which are controlled and updated
by the Master File Server. The script should accept a branch and Git
remote name to pull the files from, in addition to a directory to save the files.

If the downloaded file doesn't have the name in the first line, the script
will not attempt to import the global.

The import would be done by the OSEHRA ZGI routine if the "-i" flag is
giveen on the command line.  If the ZGI routine
doesn't exist on the server, it will attempt to install it from the same
directory as the python script.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-194
Change-Id: I2793ecd67ec82425c046efa2fbc80e36d753cbe9