Update Setup of Dox scripts and Documentation
Update the reStructuredText documentation for the OSEHRA Cross Reference
generation to account for the new parameter, -rj, which was added
recently.

Fix the URL where the RPC and HL7 messages are linked to point to the live
instance of ViViaN

Add the required tag to the -rj argument to prevent the web pages from
being generated if the file is not supplied.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-212
Change-Id: Ib2a4d7ee0064d9bb45a4f0722b1e81be59837c12