Expand information in install_information.json
Add an additional entry to the patch information which keeps track of the
package that the patch belongs to.  This is needed for an update to ViViaN
where all patches will be available to the user without selecting a
controlling package.

Since we no longer use the index as the selector, change the "value" of
each install entry to be the "Install Name" of the patch as well.

Change-Id: I600573d30d01e6a89b44395d71374aa6be13ed7f