Fix exception list usage in VistA Export
Fix the formatting of the ROUTINE_EXTRACT_EXCEPTION_LIST to be a list of
item instead of a single entry. The single entry would be used as a string
in other scripts causing the export to not function correctly.

Add the exception list to the command line usage of the VistARoutineExport
script to have the capability there as well.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-213
Change-Id: Ic37130f66157a271781b44cfdeda88212d528e9d