Add synonym mapping to ViViaN menus
Add the synonym for the VistA options which are known to the parent menu
of the option.  Generate a dictionary which maps a tuple of the parent
menu name and the child option name which points to the synonym of the
child option.  The synonym should be placed in brackets to differentiate
it from the true name of the option.

Add the paramter to the _addChildMenusToJson function so that the child
menu will be able to know what its "parent" is.

Change-Id: Id3261aab903c98d0533c4e40c32fd8e5bb0010be