Add entry point and ICR information to DOX pages
Add the capability for the WebPageGenerator to read in the OSEHRA
ICR JSON file and add links to the ViVIaN list of the ICR where it is
appropriate.

Use the parsing of the routine to generate a list of entry points which
will be used to match and display the ICR information.  Add a new section
to the page to display this information.

OSEHRA-Id: http://issues.osehra.org/browse/VIVIAN-206
Change-Id: I7e932aa1cd2ee256bb675bbe4afe36d428d71889