Add CPRS 1.0.30.75
Add the latest release of CPRS 30C (1.0.30.75) to be built by the OSEHRA
testing framework.

Add a CPRSChart.rc file which was used to generate a new version of
CPRSChart.RES.  This file is used by the BRCC32 excutable and was run as:

 $BRCC32 CPRSChart.rc

This version of CPRS requires Delphi XE3 to execute the compilation
successfully.

OSEHRA-Id: http://issues.osehra.org/browse/OAT-162
Change-Id: I2f3c4c5a9b4126fc8c33f843860435e36907a0e2