Add a 'hasLink' flag for ZZSERVERMenu entries
Add an extra flag to the JSON files of the menu which will let the file
decide if it should have a clickable link or not.

Exclude the ZZSERVER menu and all of its child menus from having a link to
click on as we generate these menus for the display only.  All other links
are valid, for now.

Change-Id: I301ed26147619ad7b650895ebddfb100c3cbab5e