Restructure setup for BCMA
Redo the order of setup for the test VistA instances such that a wards
is available when you run the TEST_VISTA_FRESH_SETUP option without having
to run the Registration RAS test

Add the capability to add a "NURS LOCATION" which is needed for the BCMA
connection.

Add the Secondary menu option and Security key needed to sign into the
BCMA GUI to both the Nurse and the Doctor.

Change-Id: Ib3f6445f7513eceec590263221aa9dc1bd1fabef