Add FOIA Patches through August 2017
Add the FOIA patches that were available on the foia-vista.osehra.org
website on August 8, 2017.

Add a Custom installer to deal with the prompts on the HT TEMPLATES
PROJECT install.

Remove the NDF*4.0*527 and NDF*4.0*528 patches as they require a patch
that OSEHRA cannot acquire in NDF*4.0*526.  427 checks for the 426 patch
as part of a pre-install routine.

Change-Id: Idd1bffa286c1393cabefc697f92fac11713d3054