Fix Documentation links
Fix the links found in the README.rst file.  Some of them used the
incorrect punctuation mark to denote a link.  Fix the SetupDox link to
point to the correctly named file.

Also, rename the file so that GitHub will render the ReStructuredText when
viewing the file in the web browser.

Change-Id: Iaca752be21af304f4189b945c22d2dc038e6c96f