Remove single field trigger for DBAComments
Remove the sole exit condition for the DBAComments section of "Date/time"
and use any other field as a reason to stop adding text to the DBA
Comments field.

This change will stop entries from being absorbed into others, ICR-10035
is missing but should exist.  Thanks to S.Habiel for the report.

Change-Id: Ie8fb296de026e8bf3aeb6cfaede1bba0aaf29864