Change Problem List category field
Fix the M-Unit test that adds a new category to the Problem List Category
file.  It tries to stuff the current date into field 1 which was changed
to be the pointer to the Category Contents.

Assume that the field to put the date into is .02 but check the DD for
file 125.11, if field 1 contains the DATE string use that instead.

SMH: Updated to use $$PATCH^XPDUTL("GMPL*2.0*49") to check for the install.
Updated to include LOCAL vs NATIONAL field

Change-Id: I13a432864194a549823ac2fe211e23fb9b5aa546