Fix display of lists in downloaded tables
Add a custom formatter for the body cells of the PDF download which will
check for '<li>' tags in the entry.  If one exists, put an asterisk at
the start of the entry.  For each entry, then remove the HTML tags
leaving a formatted list.

For CSV, ensure that the same splitting occurs, but the values are
joined by a "|" separator.

Change-Id: Idbafd6821b0dd9d8ff7770bed926042889c5602d