Add code to parse BUILD file for dependencies
Add the code to stop checking the FOIA patches for the dependency
information and pull the same information from the BUILD(#9.6) file itself.

Change the multibuild code to not use the file names, but pass an
attribute that indicates the build is a multibuild.

Ensure that the IEN of the patch in the BUILD file is passed to the
information from the INSTALL file.

Change-Id: I287a405bbc61195c24435cdc59d100dde1258a47