Update patch inquiry capability.
"Short circuit" the patch inquiry system with a call to PATCH^XPDUTIL.
If it returns "1", we do not need to check FileMan for the patch status
as it is installed.  If it doesn't, utilize the print of all patches to
determine the current install status.

These prints are necessary because a VEHU instance did not display the
patch status in the summary line.

Update the Sequence number regex to account for times where a space is not
used to separate the pound sign from the word "SEQ"

Change-Id: I3ad29400c4f65c717f996ee3ebc392674b95c62c