IPv6 issue (#26)
It turns out this code is needed after all! For some reason, the broker
sends the workstation name as the IP address. It seems that the VistA
code needs to be fixed here. In any case, I was wrong in removing it.
You hit this issue if you try to connect CPRS to an IPv6 address (e.g.
[::1]).