Start parsing split ZWR files
Start the parsing of the ZWR Global files that are split into multiple
entries.

Manually "garbage collect" during the run of the loop to ensure that
the memory usage doesn't get out of control.

Only show the first 1000 entries of a file and then place a link to
the folder of entries.