Fix % links and add entry point links
Ensure that the "%" is dropped from the routine name when generating the link
to the source code page in DOX.

Add an additional check for entry points which checks for label references
before creating the link.

TODO:  This is a little overreaching, as links to variables that happen to have
the name as an entrypoint are linked incorrectly.

  See DUZ entry point of XUP which sets the DUZ variable for this demonstrated clearly.