Use substring to replace only once for Entry points
Use the span of the REGEX match to take the area of the string
which contains the routine call or entry point and replace the text
in that smaller area. This attempts to prevent:

Malformed HTML when replacing inside already replaced lines
all instances of the string becoming links when only one is a call

Ensure regex matches are sorted in reverse to not have to recalcuate
indexes as the string is manipulated.