Add unique anchors for expanded code
Add the correct parsing for each line to add anchors to the expanded
source code display.

When creating anchors for the entry points in the expanded code display,
append an 'x' to the front of it.  This will let the anchor bypass the hidden
terse form and jump to the correct place.