Expand usage of expanded anchors
Ensure that the notation format for denoting an entry point in the
expanded display is used when generating the links to other routines.
Switch from an "x" to an underscore, "_"
For each source page, if the second character of the location.hash is
an underscore, '_', make the expanded code visible.