Update RAS test to account for whole month prompt
Update the RAS test for Registration to account for when "T-100" used in
the Disposition Summary option hits on the 1st of a month.  When that
happens, a new prompt is added before "END DATE" to ask if the
information should be calculated over the "whole month"

https://code.osehra.org/CDash/testDetails.php?test=3243957&build=44202

Say no to that new prompt.

Change-Id: I459d25d02d0cb08337f8c38a66a8f8fb37becdea