The Problem List and Scheduling RAS tests now passes in both Python 2 and 3
We cannot assume that the output of the csv DictReader is in a specific
order. Instead, sort by key to guarantee problems are added in the same
order every time.

The order of entries in the datafiles needed to be reversed. The patient
data was using ssn as the key. Instead, add a ssn column and use the row
number for the key.

https://github.com/OSEHRA/VistA/issues/117