Fix parsing of package components
Fix the parsing such that the names for each package component is properly found.
Properly check for the end of the sections as it goes.
Check for section headers before checking for "Press return..."  as one header
line in a file contained both and was throwing off the "localHandler" logic.