A couple of fixes for KIDS installer
Add timeout for connection.expect to allow system to deal with packages
that take a long time to load or start installing (e.g. Lexicon).

"Select Installation" is wrong as an indicator of bad installations.
Removed.

We no longer print "Install Completed" due to suppression of the use of
the regular home device in a previous commit. We should look for xxx
Installed now. Select Installation is also now an indicator of a good
installation, but I will probably revisit that decision in the future.