CPRS 32A Source Code with small mods
CPRS version 1.32.114.1/OR*3*539

Mods to code:
- Broker Changed to fix SID issue
- Broker had an issue with a certificate property called .SerialNum which was
not available. Replaced with a dummy value, since this code is not intended to
be used anyways outside of the VA.

This mod was not applied since the file was moved around and it's harder
now, but may do it in the future:
- Access code and Verify Code pass on cmd line

This mod is no longer needed:

- Enhanced broker logging is not needed. It's now available in the
codebase by default.

Binaries available here:

https://code.worldvista.org/files/clients/OSEHRA_VistA/CPRS/1_32_114_1/