Update KIDSBuildParser.py to recognize new KIDS "REST" node
This new node was introduced in XU*8*713 (Sep 2020), but was not
encountered till DG*5.3*112 (Feb 2024). The Python code to analyze a
build compalined that it didn't recognize "REST". This commit fixes
that.

Note that even though "REST" contains a routine, none of the routine
processing steps are set to happen since my examination of the native
KIDS code shows me that the "REST" routine actually is never saved in
VistA, only stored in a node for use in restoring the system.