Fix undefined error for self._localHandler._addVarToRoutine
In rare cases, where a label in an XINDEX output looks like an option
name (e.g. in Cache, 552P1 is a valid label, when it normally isn't),
make sure you don't crash if the localhandler variables are not
initialized, as they are checked later.

This fixes a problem found while generating Vivian for January 2025
release.